# Speed up group metadata manager test