Master issue: #830

### Motivation

The `GroupMetadataManagerTest` spend too much time to start mock pulsar and shutdown mock pulsar in each tests. See: 
https://github.com/streamnative/kop/issues/830#issuecomment-949259927

### Modifications

1. Each time we start new test method, just restart `GroupMetadataManagerTest`, instead of restart new mock pulsar.
2. Extract the common consumer

This PR speed up the `GroupMetadataManagerTest`, reduce test time from **240s** down to **20s**

**Before** : 
```text
[INFO] Running io.streamnative.pulsar.handlers.kop.coordinator.group.GroupMetadataManagerTest
[INFO] Tests run: 28, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 242.184 s - in io.streamnative.pulsar.handlers.kop.coordinator.group.GroupMetadataManagerTest
```
**After** : 
```text
[INFO] Running io.streamnative.pulsar.handlers.kop.coordinator.group.GroupMetadataManagerTest
[INFO] Tests run: 28, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 19.268 s - in io.streamnative.pulsar.handlers.kop.coordinator.group.GroupMetadataManagerTest
```